+++ spine-cpp/spine-cpp/src/spine/Atlas.cpp
@@ -192,7 +192,7 @@ struct AtlasInput {
 		line.end = index;
 		if (index != end) index++;
 		line = line.trim();
-		line.length = end - start;
+		line.length = (int)(end - start);
 		return &line;
 	}
